Se terminan las armas, se crea clase para comunicar entidades y jugador con los daños extra de las armas, además se agrega armadura nueva

+++ src/openblocks/common/entity/math/Stats.java
@@ -4,9 +4,9 @@ import net.minecraft.entity.Entity;
 
 public abstract class Stats {
 	abstract void statsAllocte();
-	public abstract float getAttack();
-	public abstract float getDefense();
-	public abstract float getMagic();
-	public abstract float getResistance();
-	public abstract float getSpeed();
+	public abstract double getAttack();
+	public abstract double getDefense();
+	public abstract double getMagic();
+	public abstract double getResistance();
+	public abstract double getSpeed();
 }
